[fix] 3-print_all.c

+++ 0x0F-variadic_functions/3-print_all.c
@@ -24,7 +24,7 @@ void print_all(const char * const format, ...)
 			printf("%c", va_arg(args, int));
 			break;
 		case 'f':
-			printf("%f", va_arg(args, float));
+			printf("%f", va_arg(args, double));
 			break;
 		case 's':
 			ptr = va_arg(args, char *);
